requirements.txt
async-lru==1.0.2
greenlet==1.1.0
psycopg2-binary==2.8.6
pyaes==1.6.1
Pyrogram==1.2.9
pyromod==1.5
PySocks==1.7.1
python-dotenv==0.17.1
SQLAlchemy==1.4.17
TgCrypto==1.2.2
flask==2.0.1
python-dotenv==0.17.1
aiohttp
dnspython

